backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.34
asyncpg==0.29.0
python-dotenv==1.0.1

